+++ packages/tizen_rpc_port/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## NEXT
 
-* Add 24 integration test cases.
+* Add 24 integration test cases to the client example.
+* Add 23 integration test cases to the server example.
 
 ## 0.1.5